Animations, footer, buttons, and error highlighter
I made some animations for the buttons and Explore mana inputs.

Began working on the footer for the page, it looks okish now.

Reworked the highlight code to replace the textarea value with two appended javascript objects, which will later be changed into Scryfall found and not found cards. That way, the user will have all errors at the top and highlighted.