urlHandler
Not much has changed. I wrote a basic script that handles the filters based on a URL, which should be useful for when the API is done, too. So, that's good.

I also changed the navbar into a more compact version, placing the clear button at the end and removing the center filters (shortcuts) in favor of a collapsed menu for all property filters.

Note: The mobile UI still needs to be fixed.